refactor: Update app according to FareProductTypeConfig changes
Mainly three tings has changed:
- Offer endpoint is now based on zone selection mode instead of
  being a field in the config.
- The field 'productSelectionTitle' on FareProductTypeConfig is
  used for the title of the product selection part in purchase
  overview.
- If ´productSelectionMode´ is ´productAlias´, then the
  ´productAlias´ on PreassignedFareProduct is used for the product
  selector. This replaces the `duration` mode as it is more
  dynamic.
