+++ src/reference-data/defaults/preassigned-fare-products.json
@@ -190,6 +190,16 @@
     "version": "ATB:Version:c7e5a0d8-78e8-46e9-a019-07348b8328db",
     "type": "period",
     "durationDays": 7,
+    "productAlias": [
+      {
+        "lang": "nob",
+        "value": "7 dager"
+      },
+      {
+        "lang": "eng",
+        "value": "7 days"
+      }
+    ],
     "distributionChannel": ["app", "web"],
     "name": {
       "lang": "nob",
@@ -218,6 +228,16 @@
     "version": "ATB:Version:12b94a4d-b347-41c9-a6c6-35bc3386d4f4",
     "type": "period",
     "durationDays": 30,
+    "productAlias": [
+      {
+        "lang": "nob",
+        "value": "30 dager"
+      },
+      {
+        "lang": "eng",
+        "value": "30 days"
+      }
+    ],
     "distributionChannel": ["app", "web"],
     "name": {
       "lang": "nob",
@@ -247,6 +267,16 @@
     "version": "",
     "type": "period",
     "durationDays": 60,
+    "productAlias": [
+      {
+        "lang": "nob",
+        "value": "60 dager"
+      },
+      {
+        "lang": "eng",
+        "value": "60 days"
+      }
+    ],
     "distributionChannel": ["app", "web"],
     "name": {
       "lang": "nob",
@@ -276,6 +306,16 @@
     "version": "",
     "type": "period",
     "durationDays": 90,
+    "productAlias": [
+      {
+        "lang": "nob",
+        "value": "90 dager"
+      },
+      {
+        "lang": "eng",
+        "value": "90 days"
+      }
+    ],
     "distributionChannel": ["app", "web"],
     "name": {
       "lang": "nob",
@@ -305,6 +345,16 @@
     "version": "ATB:Version:62736d83-3854-4c71-b9f2-d2d9b52997ab",
     "type": "period",
     "durationDays": 180,
+    "productAlias": [
+      {
+        "lang": "nob",
+        "value": "180 dager"
+      },
+      {
+        "lang": "eng",
+        "value": "180 days"
+      }
+    ],
     "distributionChannel": ["app", "web"],
     "name": {
       "lang": "nob",
