+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/Purchase/Overview/components/ProductSelection.tsx
@@ -1,40 +1,45 @@
 import {PreassignedFareProduct} from '@atb/reference-data/types';
-import {ProductSelectionMode} from '@atb/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/FareContracts/utils';
-import ProductSelectionByDuration from './ProductSelectionByDuration';
+import {FareProductTypeConfig} from '@atb/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/FareContracts/utils';
+import ProductSelectionByAlias from './ProductSelectionByAlias';
 import ProductSelectionByProducts from './ProductSelectionByProducts';
 import {StyleProp, ViewStyle} from 'react-native';
 
 type ProductSelectionProps = {
   preassignedFareProduct: PreassignedFareProduct;
-  selectionMode: ProductSelectionMode;
+  fareProductTypeConfig: FareProductTypeConfig;
   setSelectedProduct: (product: PreassignedFareProduct) => void;
   style?: StyleProp<ViewStyle>;
 };
 
 export default function ProductSelection({
   preassignedFareProduct,
-  selectionMode,
+  fareProductTypeConfig,
   setSelectedProduct,
   style,
 }: ProductSelectionProps) {
-  return (
-    <>
-      {selectionMode === 'product' && (
+  switch (fareProductTypeConfig.configuration.productSelectionMode) {
+    case 'product':
+      return (
         <ProductSelectionByProducts
           selectedProduct={preassignedFareProduct}
+          fareProductTypeConfig={fareProductTypeConfig}
           setSelectedProduct={setSelectedProduct}
           style={style}
         />
-      )}
-
-      {selectionMode === 'duration' && (
-        <ProductSelectionByDuration
+      );
+    case 'productAlias':
+    case 'duration':
+      return (
+        <ProductSelectionByAlias
           color="interactive_2"
           selectedProduct={preassignedFareProduct}
+          fareProductTypeConfig={fareProductTypeConfig}
           setSelectedProduct={setSelectedProduct}
           style={style}
         />
-      )}
-    </>
-  );
+      );
+    case 'none':
+    default:
+      return null;
+  }
 }
