+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/Purchase/Overview/components/ProductSelectionByProducts.tsx
@@ -12,28 +12,42 @@ import {StyleProp, View, ViewStyle} from 'react-native';
 import {PreassignedFareProduct} from '@atb/reference-data/types';
 import {useFirestoreConfiguration} from '@atb/configuration/FirestoreConfigurationContext';
 import * as Sections from '../../../../../../components/sections';
+import {ThemeText} from '@atb/components/text';
+import {FareProductTypeConfig} from '@atb/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/FareContracts/utils';
+import {useTextForLanguage} from '@atb/translations/utils';
+import {StyleSheet} from '@atb/theme';
 
 type ProductSelectionByProductsProps = {
   selectedProduct: PreassignedFareProduct;
+  fareProductTypeConfig: FareProductTypeConfig;
   setSelectedProduct: (product: PreassignedFareProduct) => void;
   style?: StyleProp<ViewStyle>;
 };
 
 export default function ProductSelectionByProducts({
   selectedProduct,
+  fareProductTypeConfig,
   setSelectedProduct,
   style,
 }: ProductSelectionByProductsProps) {
   const {t, language} = useTranslation();
   const {preassignedFareProducts} = useFirestoreConfiguration();
+  const styles = useStyles();
 
   const selectableProducts = preassignedFareProducts
     .filter(productIsSellableInApp)
     .filter((p) => p.type === selectedProduct.type);
   const [selected, setProduct] = useState(selectedProduct);
 
+  const title = useTextForLanguage(
+    fareProductTypeConfig.configuration.productSelectionTitle,
+  );
+
   return (
     <View style={style}>
+      <ThemeText type="body__secondary" color="secondary" style={styles.title}>
+        {title || t(PurchaseOverviewTexts.productSelection.title)}
+      </ThemeText>
       <Sections.Section>
         <Sections.RadioGroupSection<PreassignedFareProduct>
           items={selectableProducts}
@@ -69,3 +83,9 @@ export default function ProductSelectionByProducts({
     </View>
   );
 }
+
+const useStyles = StyleSheet.createThemeHook((theme) => ({
+  title: {
+    marginBottom: theme.spacings.medium,
+  },
+}));
