+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/Purchase/Overview/use-offer-state.ts
@@ -4,7 +4,6 @@ import {PreassignedFareProduct, TariffZone} from '@atb/reference-data/types';
 import {Offer, OfferPrice, searchOffers} from '@atb/ticketing';
 import {CancelToken} from 'axios';
 import {useCallback, useEffect, useMemo, useReducer} from 'react';
-import {OfferEndpoint} from '../../FareContracts/utils';
 import {UserProfileWithCount} from '../Travellers/use-user-count-state';
 
 export type UserProfileWithCountAndOffer = UserProfileWithCount & {
@@ -114,7 +113,7 @@ const initialState: OfferState = {
 };
 
 export default function useOfferState(
-  offerEndpoint: OfferEndpoint,
+  offerEndpoint: 'zones' | 'authority',
   preassignedFareProduct: PreassignedFareProduct,
   fromTariffZone: TariffZone,
   toTariffZone: TariffZone,
@@ -200,7 +199,6 @@ export default function useOfferState(
     updateOffer,
     userProfilesWithCount,
     preassignedFareProduct,
-    offerEndpoint,
     zones,
     travelDate,
   ]);
