+++ src/ticketing/api.ts
@@ -1,4 +1,3 @@
-import {OfferEndpoint} from '@atb/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/FareContracts/utils';
 import {APP_SCHEME} from '@env';
 import {AxiosRequestConfig} from 'axios';
 import {ReserveOfferRequestBody} from '.';
@@ -59,7 +58,7 @@ export type ReserveOfferWithRecurringParams = ReserveOfferParams & {
 };
 
 export async function searchOffers(
-  offerEndpoint: OfferEndpoint,
+  offerEndpoint: string,
   params: OfferSearchParams,
   opts?: AxiosRequestConfig,
 ): Promise<Offer[]> {
